refactor: removing contentlay and mdx altogether
