+++ components/quick-view.tsx
@@ -26,7 +26,7 @@ export const QuickView = ({ item }: { item: IQueryItem}) => {
         </SheetHeader>
         <ItemDetails item={item} showStatus/>
         <div className="bg-gray lg:col-span-2 p-4 flex flex-col gap-4">
-          <Comments id={item.id} status={item.status} comments={[]} canAddComments={false} />
+          <Comments id={item.id} status={item.status} canAddComments={false} />
         </div>
       </SheetContent>
     </Sheet>
