+++ package.json
@@ -18,18 +18,15 @@
     "@tanstack/react-query": "^5.59.16",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
-    "contentlayer": "^0.3.4",
     "date-fns": "^3.3.1",
     "firebase": "^11.0.1",
     "firebase-admin": "^12.7.0",
     "js-cookie": "^3.0.5",
     "lucide-react": "^0.454.0",
     "next": "^13.4.19",
-    "next-contentlayer": "^0.3.4",
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.50.1",
-    "rehype-slug": "^6.0.0",
     "tailwind-merge": "^2.2.1"
   },
   "devDependencies": {
